Create project service
